gpu network training

+++ spoken-words-recognition-api/NeuralNetwork.API/Config/GradientConfig.cs
@@ -2,9 +2,9 @@ namespace NeuralNetwork.API.Config
 {
     public class GradientConfig
     {
-        public double GradientFactor { get; set; }
-        public double WeightGradientFactor { get; set; }
-        public double BiasGradientFactor { get; set; }
-        public double InputGradientFactor { get; set; }
+        public float GradientFactor { get; set; }
+        public float WeightGradientFactor { get; set; }
+        public float BiasGradientFactor { get; set; }
+        public float InputGradientFactor { get; set; }
     }
 }
